Improved allocation logic of buttons on the WinList
When a WinList spawned more than one row, the buttons would be placed on a
grid. As sizing of the buttons was relative but limited to the confinement
of the grid, often large amounts of free space would be left in one row
while sometimes needlessly cutting short a button on the other row.

This change tries to make better use of the available space by looking at
the available space per row instead of using a rigid column structure.

The number of buttons per row is also more dynamic and based on the amount
of space available.